bug fixing.

+++ app/src/main/java/gridwatch/plugwatch/utilities/Reboot.java
@@ -4,6 +4,8 @@ import android.app.IntentService;
 import android.content.Intent;
 import android.util.Log;
 
+import gridwatch.plugwatch.configs.AppConfig;
+
 /**
  * Created by nklugman on 11/17/16.
  */
@@ -22,17 +24,19 @@ public class Reboot extends IntentService {
     public void do_reboot() {
         Log.e("restart", "doing reboot");
 
-        try {
-            Process proc = Runtime.getRuntime()
-                    .exec(new String[]{"su", "-c", "reboot"});
-            proc.waitFor();
-            Runtime.getRuntime().exec(new String[]{"/system/bin/su", "-c", "reboot"});
-        } catch (Exception ex) {
-            ex.printStackTrace();
-            if (ex.getCause().getMessage().toString().equals("Permission denied")) {
+        if (!AppConfig.TURN_REBOOT_OFF) {
+            try {
+                Process proc = Runtime.getRuntime()
+                        .exec(new String[]{"su", "-c", "reboot"});
+                proc.waitFor();
+                Runtime.getRuntime().exec(new String[]{"/system/bin/su", "-c", "reboot"});
+            } catch (Exception ex) {
+                ex.printStackTrace();
+                if (ex.getCause().getMessage().toString().equals("Permission denied")) {
+
+                }
 
             }
-
         }
     }
